Data/python/requirements.txt
hyperlpr==0.0.1
mysql-connector-python==8.0.16
